LIBDRUM-743. Changed autoCommit time for Solr "statistics" core
Modified the Solr "statistics" core configuration to perform an
autoCommit every 10 seconds, instead of every 15 minutes. This is
intended to provide near-immediate updates to the "No. of downloads"
counts on the item pages.

Added instructions fot "Drum7DockerDevelopmentEnvironment.md" on
setting up the local development environment with the "GeoIP"
database, which is necessary to test the file download counts.

https://umd-dit.atlassian.net/browse/LIBDRUM-743