LIBDRUM-753. Store DOIs using the "URL" form, instead of "doi" form
In the stock DSpace 6 implementation, DOIs are placed in the
"dc.identifier.doi" field, and in the database are stored in the
"URL" form (i.e., as "https://doi.org/10.13016/gre8-qkiy").

In LIBDRUM-562, the DSpace 6 code was modified to store DOIs in the
"dc.identifier" field, and to store them in the
"doi" form (i.e., "doi:10.13016/gre8-qkiy").

While working on the DSpace 7 migration, it was noticed that on the
on the production server, DOIs are being placed in the
"dc.identifier" field, but are being stored in the database using the
"URL" form. Of the 14,005 DOIs in the production database, only 14 were
in the "doi" form, all the other 13,991 entries are in the "URL" form.
It is not clear why the DSpace 6 change to store DOIs in the "doi"
format was ineffective.

This commit restores the stock DSpace functionality of storing DOIs
in the "url" form, which is most consistent with what is currently in
the database.

https://umd-dit.atlassian.net/browse/LIBDRUM-753