LIBDRUM-757. Incorporate fix from DSpace 7.5
The DSpace 7.4 error where a search query with quotes fails when
"Browsing by Title" was fixed in DSpace 7.5 with the following commmit:

https://github.com/DSpace/DSpace/commit/b0afb15efa6ad962ff2447fdbaf78c23813758ec

This change incorporates that fix into our DSpace 7.4 codebase.

This class can be removed when the application has been upgraded to
DSpace 7.5.

Note: This change does not preserve the DSpace 6 behavior relating to
quote handling, but does align with DSpace 7.5 behavior.

https://umd-dit.atlassian.net/browse/LIBDRUM-757