LIBDRUM-667. Restored Checkstyle checking in Jenkinfile
Removed the "-Dcheckstyle.skip=true" flag when building code using
Jenkins. This ensures that the Checkstyle checks will be run, and
any warnings flagged.

Added "recordIssues" so that Checkstyle results would be available
in Jenkins.

Configured "recordIssues" so that the hundreds of INFO-level
Checkstyle warnings about TODO comments from the stock DSpace
code are not included in the health check count, or in the
number of violations sent in the email.

https://umd-dit.atlassian.net/browse/LIBDRUM-667