LIBDRUM-769. Added caching for LDAP queries
When impersonating users, an LDAP query may need to be performed for
each HTTP request, so a simple in-memory expiring cache, using the Apace
Commons "PassiveExpiringMap" class was added.

By default, this will store an LDAP result for 5 minutes before
the entry in the cache expires.

Added value for "drum.ldap.cacheTimeout" to "local.cfg.EXAMPLE".

Added unit tests to verify cache behavior.

https://umd-dit.atlassian.net/browse/LIBDRUM-769