Switch all DAOs to take Session rather than Context.
This frees us to use DAOs with multiple Sessions per "request".

A consequence of this change is that DAOs may not call model object
Services.  I consider doing that a layering violation, so I treat this
as a feature.  Only RelationshipDAOImpl reqired changes, so see that
for the sort of changes required (which were minimal).