DS-8935. webui.browse.link CrossLinks - Fix for multiple exact matches
Fixes #8935 when multiple exact match "webui.browse.link" configuration
entries are present that point to different indexes.

Modified the code to return the index associated with the given
metadata (which is used as the key in the hash map), instead of the key
from the keySet (which may not actually be the metadata value being
searched for).

https://github.com/DSpace/DSpace/issues/8935
(cherry picked from commit b846c53baaeae1e19fbbafa3dc7ca724bcaf32c1)