LIBDRUM-812. Placed stock DSpace 7.6 files into dspace-api test config
Downloaded stock DSpace 7.6 configuration files that were modified
in the "config/dspace" directory into the
"dspace-api/src/test/data/dspaceFolder/config" directory, so that the
"dspace-api" tests would have the expected configuration (and so the
tests could run and pass succesfully).

This is needed because the "dspace/config" files have been customized
to refer to UMD-added classes that are not available on the classpath
used for the "dspace-api" tests.

See the "XML File Changes" section in
"dspace/docs/DrumConfigurationCustomization.md" for more information.

https://umd-dit.atlassian.net/browse/LIBDRUM-812