LIBDRUM-815. Modify local development environment to use HTTPS
Modified local development environment to use HTTPS, in order to
restore CAS authentication.

In September 2023, DIT changed the CAS application firewall to reject
with a "403 Forbidden" HTTP status code any CAS requests with a
"service" HTTP parameter of "localhost" or "127.0.0.1".

This requires the local development environment to use a hostname other
than "localhost" or "127.0.0.1", which in turn requires:

* Use the HTTPS protocol for all interactions with CAS
* Use a "full" hostname
* Use the "mkcert" application to register a certificate authority
* Have "drum-local.lib.umd.edu" and "api.drum.lib.umd.edu" hostnames
  added to the "/etc/hosts" file
* An Nginx container in the "docker-compose.yml" file for HTTPS handling

See the changes to "dspace/docs/CASAuthentication.md" for additional
explanation of the necessity for these changes.

https://umd-dit.atlassian.net/browse/LIBDRUM-815