LIBDRUM-819. Added "script-mail-wrapper" script for emailing script logs
Added the "dspace/bin/script-mail-wrapper", that "wraps" a script so
that the script is run with the resulting log file sent to a specified
email address via the "mail" script.

This is intended to replace the need for extensive redirection in the
Kubernetes "crontab.txt" file that "streams" the log output to the SMTP
server (as used by the "load-etd-nightly" cron entry).

https://umd-dit.atlassian.net/browse/LIBDRUM-819