Remove the is_archived_condition from the doi-filter
In the conversation of #9271 kshepherd states that he had issues to
register DOIs when the doi-filter is usint the is_archived_condition. I
was able to reproduce these issues. This removes the reference to that
condition from the filter. The item is not archived when the
InstallItemService is trying to apply a DOI to it.