LIBDRUM-814. Remove need to "pin" Postgres to 13.10
The need to "pin" the Postgres version to 13.10 appears to have been
caused by the fix to
[CVE-2023-2454](https://www.postgresql.org/support/security/CVE-2023-2454/).

The reason why the fix breaks this functionality is unclear (it may be
due to the database initialization process running as the "postgres"
user, and the use of the "pgcrypto" extension).

After numerous attempts to get the database initialization working as it
was prior CVE-2023-2454 were unsuccessful, decided to take advantage of
the automatic “*.sh” scripts handling of the "postgres" Docker image
to do the following:

1) Instead of using a plain SQL dump, use the Postgres “custom” dump.

2) Since Postgres “custom” dumps are not automatically, processed,
created a “postgres-init/drum_pg_restore.sh” Bash script to use
“pg_restore” to process any files with a “.dump” extension in the
“/docker-entrypoint-initdb.d” directory.

This preserves most of the convenience of the prior method.

Made the following additional changes:

* Removed the “dspace/docs/DrumDBRestoreFromDSpace6.md”, as we no longer
  have any servers running DSpace 6.

* Updated the following documentation:
    * postgres-init/README.md
    * dspace/docs/DrumDBRestore.md
    * dspace/docs/Drum7DockerDevelopmentEnvironment.md

* Updated “.gitignore” to track the “postgres-init/drum_pg_restore.sh”
  Bash script

* Removed the UMD customization pinning the Postgres version to “13.10”
  in dspace/src/main/docker/dspace-postgres-pgcrypto/Dockerfile

https://umd-dit.atlassian.net/browse/LIBDRUM-814