LIBDRUM-876. Make "Equitable Access" field dynamic based on "Type"
For the default submission form, created a "Submission Type" step
("umdType") and associated form to enable the "Equitable Access" field to
be dynamically displayed/hidden based on whether "Article" was chosen in
the "Type" dropdown.

This also required creating a separate "Submission Type" step
("mhheaType") and associated form to display the "Type" field without
the "Equitable Access" field, since MHHEA submissions are never use the
"Equitable Access" mapping.

Removed the "equitableAccessSubmission" step and associated form, as it
is no longer used.

https://umd-dit.atlassian.net/browse/LIBDRUM-876