LIBDRUM-876. Modified Author/Advisor label for desired placeholder text
DSpace does not have the ability to set the placeholder text on form
fields. Instead, it simply reuses the label for the field.

Terry Owen requested that the "Author" and "Advisor" placeholder text
be changed to "Last Name, First Name", in order to prompt users to enter
names in that order.

Since it is not currently possible set the placeholder text separate
from the label, changed the labels to:

* Author (Last Name, First Name)
* Advisor (Last Name, First Name)

which also changes the placeholder text.

https://umd-dit.atlassian.net/browse/LIBDRUM-876