fix(submission): Submission scope naming fixed
According to the documentation, the value for the <visibility> property is 'submission' in the 'submission-forms.xml'. Without this change, an empty input field will never be marked as an error, even if the field is marked as 'required'.

(cherry picked from commit 02f52c7d5c245eaae09dbd636fc8c7935d7cc242)