LIBDRUM-892. Update parent POM version for "server-boot" module
Updated the parent POM version for the "server-boot" module to point to
the correct parent (i.e., DRUM, instead of DSpace).

This was missed in the initial pass, because the "server-boot" module
was not present in DSpace 8.

Checkstyle fixes to imports, so that code will pass the Checkstyle task
when running the Maven build.

https://umd-dit.atlassian.net/browse/LIBDRUM-892