LIBDRUM-892. Refactored Dockerfiles for external Tomcat
In DSpace 8, the stock DSpace Dockerfiles were modified to use a
Spring Boot "embedded Tomcat" runnable jar for running the DSpace
back-end.

This method of running ignores the UMD customizations in the
"dspace/modules/additions" and "dspace/modules/server" directories
(see DSpace Issue 9987).

In this pull request, updated the Dockerfiles (and POM files)
to use the "tomcat" Docker image for running the application,
as in DSpace 7.6.2. This ensures that the ensure that UM
 customizations in the "dspace/modules/additions" and
"dspace/modules/server" directory are available.

https://umd-dit.atlassian.net/browse/LIBDRUM-892