LIBDRUM-892. Refactored Dockerfiles and POMs for embedded Tomcat
Modified the Dockerfiles and POMs to be more similar to the stock
DSpace 8 Dockerfiles, which use a Spring Boot "embedded Tomcat" runnable
jar for running the DSpace back-end.

This commit includes customizations to the POM files sent to DSpace
in https://github.com/DSpace/DSpace/pull/10043, so that customizations
in the "additions" and "server" modules will be incorporated into the
embedded Tomcat build.

Updated "DrumConfigurationCustomization" to provide links to the
"fix" for the embedded Tomcat server.

https://umd-dit.atlassian.net/browse/LIBDRUM-892