Add a curation task to copy fields
This Curation task can copy multiple source fields into one field. It
can be used multiple times in parallel for different fields. It can
remove the fields it copied or delete the target field before copying
the source fields. A concatenator can be configured, for example if the
values of the source fields should be sparated by a comma and a space.