LIBDRUM-907. Fix ETD loader resource policy for embargoed bitstreams
In DSpace 7.6.2 changes were made to the
“dspace-api/src/main/java/org/dspace/authorize/ResourcePolicyServiceImpl.java”
class, in which the “create” method was modified to include “EPerson”
and “Group” parameters.

When originally updated for DSpace 7.6.2, the DRUM code set these
parameters were set to “null” to satisfy the compiler (as the
proper group was subsequently set using a “setGroup” method), but this
ran afoul of the parameter checking in the “create” method, resulting
in the error.

Corrected the issue by supplying the proper Group to the “create”
method, and removed the subsequent “setGroup” calls, as they were no
longer needed.

https://umd-dit.atlassian.net/browse/LIBDRUM-907