LIBDRUM-915. Updated docker-compose.yml to support VS Code debugging
In DSpace 7.6.2, the back-end was run using the Tomcat “catalina.sh”
script, which automatically included the “JPDA_OPTS” environment
variable, if present in the shell.

With the switch to using a Spring Boot embedded Tomcat server, the
Tomcat server is started by a “java”  command, instead of the
“catalina.sh” script.

Modified the “docker-compose.yml“ to use the “JPDA_OPTS” environment
variable when starting the DRUM back-end, i.e. changed:

```yaml
    entrypoint:
      ...
      java -jar /dspace/webapps/server-boot.jar --dspace.dir=/dspace
```

to

```yaml
    entrypoint:
      ...
      java $${JPDA_OPTS} -jar /dspace/webapps/server-boot.jar --dspace.dir=/dspace
```

The use of the ${{...} was suggested by https://stackoverflow.com/a/65564161

Updated the “DockerDevelopmentEnvironment.md” document to reference the
“JPDA_OPTS” environment variable, instead of the non-existent
“JAVA_TOOL_OPTIONS” environment variable.

https://umd-dit.atlassian.net/browse/LIBDRUM-915