LIBDRUM-916. Purge DOIs with no associated bitstream before registration
Modifies the "register-all" command in DOIOrganiser to delete any
DOIs in a "TO_BE_REGISTERED" state that does not have an associated
item ("dspace_object" in the "doi" Postgres table is null).

This is necessary because DRUM generates a DOI when an item is created,
but if the item is deleted before the DOI is registered with DataCite,
a NullPointerException is thrown.

Added "DOIOrganiserTest.java" to verify operation of the
"purgeEmptyDOIsWithStatus" method.

Added information about this change to "DrumDOI.md" documentation.

https://umd-dit.atlassian.net/browse/LIBDRUM-916