LIBDEVOPS-1754. Configure AWS S3 assetstore
In "bitstore.xml" added an additional assetstore (assetstore "5") as
an AWS S3 assetstore.

In order to use it for new items, DRUM must be configured to use
assetstore "5" as the primary assetstore.

Note: Since the "local.cfg.EXAMPLE" file defaults to using
assetstore "0", the local development environment should be unaffected
by this change.

https://umd-dit.atlassian.net/browse/LIBDEVOPS-1754