LIBDRUM-937. Modify bitstore migration to skip missing files
Modified the stock DSpace bitstore migration code to skip bitstreams
that are not present in the assetstore.

While not needed for production, the ability to skip bitstreams is
needed for the "sandbox", "test" and "qa" DRUM migrations, as these
instances have production database snapshots, but not all of the
production bitstreams (the stock DSpace migration code fails with an
error if any bitstream is missing).

The changes in this commit can be removed when the AWS S3 migration
is complete.

https://umd-dit.atlassian.net/browse/LIBDRUM-937