LIBDRUM-940. Added "openssh-client" to DRUM Docker image
Modified the “Dockerfile” file to include the “openssh-client” as part
of the apt-get update command used in generating the Docker image.

Also updated the “Dockerfile.dev” file for consistency with the
“Dockerfile”.

The “Dockerfile.dev-additions” file, used in local development was not
modified, because it did not seem to be necessary (the packages can be
added as needed by developers).

https://umd-dit.atlassian.net/browse/LIBDRUM-940