LIBDRUM-946. Fix checkpointing for bitstore migration
Fixes the checkpointing for bitstore migration by actually committing
the changes to the database.

This change was provided to DSpace as DSpace Pull Request 10489. The
"UMD Customization" markers can be removed once this pull request is
accepted by DSpace.

Replacing "dispatchEvents" with "commit" as testing has shown that this
is necessary to ensure that the status of bitstreams is properly
updated when the bitstore migration is interrupted.

The provided integration tests fail if bitstreams successfully migrated
before the interruption are not properly recordedd in the database as
being in the destination assetstore.

https://umd-dit.atlassian.net/browse/LIBDRUM-946