LIBDRUM-947. Updated Spring Boot embedded Tomcat for CVE-2025-24813
Updated Spring Boot embedded Tomcat version to v10.1.39 to mitigate
CVE-2025-24813.

This change is based on the "patch.diff" file provided on the
"dspace-tech" mailing list in a March 18, 2025 email from
mrwerdo331@gmail.com.

This customization can be removed when DRUM is upgraded to a DSpace
version that uses a Tomcat no longer affected by the CVE.

https://umd-dit.atlassian.net/browse/LIBDRUM-947