LIBDRUM-957. Updates to Dockerfile configuration for DRUM
Fixed capitalization of "AS" directive, and modified "ENV" directives to
use "=" syntax to fix Docker warnings.

In "Docker.dev-additions", remove the "dspace-installer" directory
commands, because it is not necessary. Also removed the "apt-get update"
that installed "host" as it does not seem necessary (handle server is
not typically used in local development), and was slowing
down the build.

Updated the "DockerDevelopmentEnvironment.md" installation instructions.

https://umd-dit.atlassian.net/browse/LIBDRUM-957