LIBDRUM-957. Update Spring Boot to mitigate CVE-2025-24813
In DSpace 8.0, mitigating CVE-2025-24813 required specifying a Tomcat
version with the fix.

Based on a March 18, 2025 email from Tim Donohue to the "dspace-tech"
mailing list, in DSpace 8.1, the Spring Boot version can simply be
updated to 3.4.3 to fix the CVE.

It turned out to also be necessary to bump the versions of
"spring"and "spring-security" to be compatible with spring-boot.

Removed older customizations for DSpace 8.0.

https://umd-dit.atlassian.net/browse/LIBDRUM-957