LIBDRUM-957. Rolled back use of "dspace-8_x" Docker tag
In DSpace 8.1, the stock DSpace files started using the "dspace-8_x" tag
instead of "latest" as the default Docker image tag.

In the initial merge, this change was accepted for DRUM, including
changing the name of the "drum-dependencies-8_x" image to
"drum-dependencies".

This turned out to be an issue when building Docker images in
Kubernetes, because the "Dockerfile.dependencies" Docker image needs
to be pushed to the Nexus in order for it to be available for subsequent
builds (the Kubernetes "build" context does not seem to have a Docker
cache store that persists between invocations).

The Nexus is set up to reject the push of any image with a tag that
already exists -- except for the "latest" tag. If we used the
"dspace-8_x" tag, this means that the "drum-dependencies" image could
only be created once.

The "Dockerfile.dependencies" Docker image uses the "pom.xml" files of
the current DRUM checkout, so clearly it must be able to be updated as
DRUM changes. Therefore, the name of the "Dockerfile.dependencies"
Docker image name was kept as "drum-dependencies-8_x" (as it was in
DRUM DSpace 8.0), and the "latest" tag remains as the default tag.

Updated all references to the "drum-dependencies" image. Also updated
the default "dspace-8_x" tag used by other images to "latest" for the
same reason as above.

https://umd-dit.atlassian.net/browse/LIBDRUM-957