LIBDRUM-962. UmdExtendedJsonAccessLogValve implementation
Subclass of stock Tomcat JsonAccessLogValve that allows an additional
JSON attribute to be appended to the log entries.

The additional attribute is appended by adding a "#<KEY>:<VALUE>#"
pattern to the log's "pattern" configuration, where <KEY> is the
JSON attribute key, and <VALUE> is the JSON attribute value.

This is intended to provide a mechanism for identifying the log file the
entry was intended for.

https://umd-dit.atlassian.net/browse/LIBDRUM-962