LIBDRUM-964. Modified "script-mail-wrapper" to support JSON-based logs
It is preferred to use JSON-formatted logs for Kubernetes, as it makes
the logs easier to search/filter in Splunk.

Modified the "script-mail-wrapper" to handle JSON-formatted logs by
extracting only the "message" field of non-DEBUG log messages to send
in the email. For Kubernetes, this preserves the existing functionality
of having non-DEBUG messages in the email sent to the mailing list,
while additional DEBUG-level information is sent to the standard out
and to Splunk.

Note that with these changes, logging in the local development
environment and Kubernetes are different -- the local development
environments uses a text-based log, instead of JSON-formatted, for
simplicity, and does not log at a DEBUG level be default.

The intention is to at least preserve similarity of the emails between
the local development environment and Kubernetes.

https://umd-dit.atlassian.net/browse/LIBDRUM-964