LIBDRUM-972. Remove EtdLoader MARC Bash scripts
Removed the "etd2marc-mail" and "etd2marc-transfer" scripts because
based on a May 29, 2025 discussion with Ben Wallberg in the "Dspace"
WebEx channel, this is legacy functionality that is no longer needed.

Also removed the "drum.etdloader.transfermarc" property, which is no
longer needed from the "load-etd-nightly" script, and the
"local.cfg.EXAMPLE".

Removing these scripts at this time because the use of the
"dspace dprop" command in the "load-etd-nightly" Bash script was
failing in Kubernetes because the logs from the console were being
included in the output returned to the "transfermarc" variable,
causing the script to fail.

https://umd-dit.atlassian.net/browse/LIBDRUM-972