LIBDRUM-972. Remove MARC and CSV genereation for EtdLoader
Removed the functionality to create MARC and CSV files from the
EtdLoader class and related scripts.

The MARC functionality is definitely legacy, and was removed based on a
May 29, 2025 conversation with Ben Wallberg in the "DSpace" WebEx
channel.

The CSV functionality was removed because it is not clear that is used
used (it is not absolutely certain that this is legacy functionality,
but is definitely suspected to be).

Removed the "etd2marc.xsl" file, because it is presumably no longer
needed.

Removed the command-line arguments related to the MARC and CSV files
from the "load-etd" and "load-etd-nightly" scripts, because they are
no longer needed.

https://umd-dit.atlassian.net/browse/LIBDRUM-972