PATCH add now supports adding/replacing multiple values with additional index checks
To better comply with RFC 6902.
If multiple values (without index) are given, these are added (and earlier values replaced, if any) to given metadata field.
When a single value is given, index is checked (number format and index bounds) and value is appended to position based on index (not only 0 or -1).
It is now considered an error if both index and list of multiple values are given.
Accommodates changes in DSpace 10.