Use canonical path for bitstream path/assetstore path comparisons
Replaced `baseDir.getAbsolutePath()` with `baseDir.getCanonicalPath()`
in the bitstream path check.

This is necessary because the "normzalizedPath" is generated via the
"getCanonicalFile" method. This method follows symbolic links, unlike
the "getAbsolutePath" method, which does not.

Therefore, when comparing the "normalizedPath" to the "baseDir", it
is necessary to use "getCanonicalPath" so that any symbolic links are
resolved in the same way.

This issue was uncovered by running the
"org.dspace.storage.bitstore.BitstreamStorageServiceImplIT" integration
tests on macOS, in which a temporary file asset store is used. On macOS
the temporary files are stored in "/var" which is a symbolic link to
"/private/var".

(cherry picked from commit 8f8cb52658954234c05e1ea71e43375f080b7d0f)