Curator does not support items without handles
The curation task CreateMissingIdentifiers should create all identifiers
an item misses. Unfortunately it will not be started, if the item does
not have a handle, as the curator falls back to handles, even if a uuid
is given as identifier.

This PR changes the Curator to check if an identifier is a uuid if no
handle was found. I added a test to the CreateMissingIdentifiersIT. It
would probably be better to add a test to the Curator, but the
CreateMissingIdentifier task seems to be the most likely target for such
a scenario.